Create Python_script_Initial_Outline
This Python script utilizes the Pandas library to conduct a quick overview of the dataset. It begins by extracting the column titles and determining the number of data rows, providing an initial understanding of the dataset's structure. Subsequently, it iterates through specified columns to identify distinct values and their respective counts, offering insights into the dataset's content. Additionally, the script calculates the maximum, minimum, and average values for 'QUANTITYORDERED' and 'SALES', aiding in determining appropriate data types and sizes for efficient storage. Overall, the script efficiently performs data exploration on the CSV dataset.